[rl] Route RL CI with pytest markers
